move all PG examples and update to use new driver approach
