+++ examples/pg/02.simple-insert.js
@@ -5,13 +5,16 @@
  * the value should be single quoted.  Number values should be included without the single quotes.
  */
 
-import { getPool } from './utils.js'
-import  { TransactionManager } from '../index.js'
-import { Root } from '../src/root.js'
+import { getPool } from '../utils.js'
+import  { TransactionManager } from '../../index.js'
+import { Root } from '../../src/root.js'
 import 'dotenv/config'
+import { PgDataDriver } from '../../src/data-drivers/pg-driver.js'
 
 const pool = getPool()
 const client = await pool.connect()
+
+const driver = new PgDataDriver( client )
 
 const commands = [
   { 
@@ -31,7 +34,7 @@ const commands = [
 ]
 
 const r = new Root( commands )
-const t = new TransactionManager( client, r )
+const t = new TransactionManager( driver, r )
 const response = await t.executeTransaction( { id: 1, status: "active", name1 : "build new widgets", name2 : "build more sprokets"  } )
 
 console.log( '== 02.simple-insert results ===================================================')
