+++ examples/pg/04.fetch-all.js
@@ -1,15 +1,18 @@
-import { getPool } from './utils.js'
-import  { TransactionManager } from '../index.js'
-import { Root } from '../src/root.js'
+import { getPool } from '../utils.js'
+import  { TransactionManager } from '../../index.js'
+import { Root } from '../../src/root.js'
 import 'dotenv/config'
+import { PgDataDriver } from '../../src/data-drivers/pg-driver.js'
 
 const pool = getPool()
 const client = await pool.connect()
 
+const driver = new PgDataDriver( client )
+
 const r = new Root( )
 r.addCommand({ name: 'fetch-all', sql: 'SELECT * FROM projects;' })
 
-const t = new TransactionManager( client, r )
+const t = new TransactionManager( driver, r )
 
 const response = await t.executeTransaction( )
