+++ examples/pg/06.sql-command-schema-failure.js
@@ -1,10 +1,13 @@
-import { getPool } from './utils.js'
-import  { TransactionManager } from '../index.js'
-import { Root } from '../src/root.js'
+import { getPool } from '../utils.js'
+import  { TransactionManager } from '../../index.js'
+import { Root } from '../../src/root.js'
 import 'dotenv/config'
+import { PgDataDriver } from '../../src/data-drivers/pg-driver.js'
 
 const pool = getPool()
 const client = await pool.connect()
+
+const driver = new PgDataDriver( client )
 
 // should fail due to param property spelled incorrectly and actual params property not being an array
 const r = new Root( [
@@ -15,7 +18,7 @@ const r = new Root( [
   } 
 ])
 
-const t = new TransactionManager( client,r )
+const t = new TransactionManager( driver, r )
 const results = await t.executeTransaction( )
 
 console.log( '== 04.fetch-all results =======================================================')
