+++ examples/pg/08.transform-contacts.js
@@ -3,13 +3,16 @@
  * The output needs to have the phones in one property instead of 4 separate properties.  The last logic step transforms the output into the necessary shape.
  */
 
-import { getPool } from './utils.js'
-import  { TransactionManager } from '../index.js'
-import { Root } from '../src/root.js'
+import { getPool } from '../utils.js'
+import  { TransactionManager } from '../../index.js'
+import { Root } from '../../src/root.js'
 import 'dotenv/config'
+import { PgDataDriver } from '../../src/data-drivers/pg-driver.js'
 
 const pool = getPool()
 const client = await pool.connect()
+
+const driver = new PgDataDriver( client )
 
 const commands = [
   { 
@@ -38,7 +41,7 @@ const commands = [
 ]
 
 const r = new Root( commands )
-const t = new TransactionManager( client, r )
+const t = new TransactionManager( driver, r )
 const response = await t.executeTransaction( )
 
 console.log( '== 02.simple-insert results ===================================================')
